Commented out dome conflicts resolution tests

+++ src/LiteCore/src/LiteCore.Shared/Interop/C4Base_defs.cs
@@ -25,7 +25,7 @@ using LiteCore.Util;
 
 namespace LiteCore.Interop
 {
-    internal enum C4ErrorDomain : uint
+    internal enum C4ErrorDomain : byte
     {
         LiteCoreDomain = 1,
         POSIXDomain,
